- rename NodeProvider, ContextNode - seperate NodeProvider, ContextNode - try add treeViewController - add treeItem onClick function
