+++ dbux-code/package.json
@@ -57,6 +57,10 @@
 			{
 				"command": "dbuxEvents.deleteEntry",
 				"title": "Delete"
+			},
+			{
+				"command": "dbuxEvents.itemClick",
+				"title": "Clicked"
 			}
 		],
 		"menus": {
