+++ dbux-code/src/codeControl/codeNav.js
@@ -12,8 +12,8 @@ const { log, debug, warn, error: logError } = newLogger('CodeNav');
  * @param {vscode.Uri} URI (new vscode.Uri.file(FILEPATH))
  * @param {vscode.Position} position (new vscode.Position(LINE, CHARACTER))
  */
-export function navToCode(URI, position){
-  log(`Called navToCode with params { URI = ${URI}, position = ${JSON.stringify(position)} }`)
+export function navToCode(URI: Uri, position: Position){
+  debug(`Called navToCode with params { URI = ${URI}, position = ${JSON.stringify(position)} }`)
   window.showTextDocument(URI).then( editor => {
     const range = editor.document.lineAt(position._line).range;
     editor.selection =  new Selection(range.start, range.end);
