+++ dbux-code/src/codeControl/index.js
@@ -1,7 +1,6 @@
 import { newLogger } from 'dbux-common/src/log/logger';
 import { initCodeDeco } from './codeDeco';
 
-
 const { log, debug, warn, error: logError } = newLogger('CodeControl');
 
 export function initCodeControl(context) {
