+++ dbux-code/src/treeView/treeData.js
@@ -1,23 +1,19 @@
 import { newLogger } from 'dbux-common/src/log/logger';
-import { getCodePositionFromLoc } from '../util/codeUtil';
 import ExecutionContextType from 'dbux-common/src/core/constants/ExecutionContextType';
-import path from 'path';
-import { 
-    TreeItem,
-    EventEmitter,
-    TreeItemCollapsibleState as CollapsibleState 
-} from 'vscode';
+import ContextNode from './ContextNode';
 
 const { log, debug, warn, error: logError } = newLogger('TreeData');
 
-export class EventNodeProvider {
+export class ContextNodeProvider {
 
-    constructor(dataProvider) {
+    constructor(dataProvider, onChangeEventEmitter) {
         this.dataProvider = dataProvider;
         this.nodesByContext = [];
         this.rootNodes = [];
-        this._onDidChangeTreeData = new EventEmitter();
-        this.onDidChangeTreeData = this._onDidChangeTreeData.event;
+        this.onChangeEventEmitter = onChangeEventEmitter;
+        this.onDidChangeTreeData = onChangeEventEmitter.event;
+
+        this.dataProvider.onData('executionContexts', this.update);
     }
 
     contextToNode = (context: ExecutionContext) => {
@@ -28,7 +24,7 @@ export class EventNodeProvider {
         
         const staticContext = this.dataProvider.collections.staticContexts.getById(staticContextId);
         debug('staticContext =', staticContext)
-        const { programId, name, displayName, loc } = staticContext
+        const { programId, displayName, loc } = staticContext
         
         const programContext = this.dataProvider.collections.staticProgramContexts.getById(programId);
         debug('programContext =', programContext)
@@ -46,15 +42,15 @@ export class EventNodeProvider {
             stackDepth,
             contextId,
             parentContextId,
-            parentNode
+            parentNode,
+            this
         );
 
         return newNode;
 
     }
     refresh = () => {
-        this._onDidChangeTreeData.fire();
-        log(this.nodesByContext)
+        this.onChangeEventEmitter.fire();
     }
     update = (newContextData) => {
         debug('Called update function.');
@@ -89,68 +85,6 @@ export class EventNodeProvider {
         }
     }
     getParent = (node: ContextNode) => {
-        return node.parentNode
+        return node.parentNode;
     }
-}
-
-export class ContextNode extends TreeItem {
-
-	constructor(
-        displayName,
-        typeName,
-        fileName,
-        filePath,
-        location,
-        depth,
-        contextId,
-        parentContextId = null,
-        parentNode = null
-    ) {
-        // set label
-        super(`${displayName} [${typeName}]`);
-
-        // node data
-        this.displayName = displayName;
-        this.typeName = typeName;
-        this.fileName = fileName;
-        this.filePath = filePath;
-        this.location = location;
-        this.depth = depth;
-        this.contextId = contextId;
-        this.parentContextId = parentContextId;
-        this.parentNode = parentNode;
-
-        // treeItem data
-        this.children = [];
-        this.description = `@${fileName}:${location.start.line}:${location.start.column}`;
-        this.collapsibleState = CollapsibleState.None;
-        this.command = 'dbuxExtension.showMsg';
-        this.contextValue = 'event';
-		this.iconPath = {
-			light: path.join(__filename, '..', '..', 'resources', 'light', 'dependency.svg'),
-			dark: path.join(__filename, '..', '..', 'resources', 'dark', 'dependency.svg')
-        };
-	}
-
-    pushChild(child){
-        this.children.push(child);
-        this.collapsibleState = CollapsibleState.Collapsed;
-    }
-
-    get(){
-        return {
-            label: this.label,
-            position: this.position,
-            collapsibleState: this.collapsibleState,
-            command: this.command, 
-            parent: this.parent.label,
-            children: this.children.map(e => e.label),
-            contextValue: this.contextValue
-        }
-    }
-
-	get tooltip() {
-		return `at ${this.fileName}(tooltip)`;
-	}
-
 }
